BGP-236 feat-fe: 채팅 메세지 발신자 프로필이미지 표시
