+++ frontend/src/type/api/chat.d.ts
@@ -1,0 +1,8 @@
+export interface Message {
+  chatId: number;
+  isMyMessage: boolean;
+  stringContent: string;
+  chatName: string;
+  createdAt: string;
+  profileImage: string;
+}
